Big Update with Randomization
Fixed bug not properly scaling the Rows and Columns of the Maze. Also created a class that instantiates the player prefab on a random location on the maze. This class also stores the centre of each maze cell ising the x,y,z coordinates. Each new generated maze also now has a randomization element that allows for a range of sizes in the rows and columns.